Port Oracle Arel visitor tests to RSpec
